Add screenshot pages for exercise and session editors
