+++ LiftLog.Ui/Pages/Screenshot/ScreenshotCollectorPage.razor.cs
@@ -1,10 +1,6 @@
 #if DEBUG
 using Fluxor;
-using LiftLog.Lib;
-using LiftLog.Lib.Models;
 using LiftLog.Ui.Services;
-using LiftLog.Ui.Store.App;
-using LiftLog.Ui.Store.CurrentSession;
 using LiftLog.Ui.Store.Program;
 using LiftLog.Ui.Store.Settings;
 using Microsoft.AspNetCore.Components;
@@ -24,7 +20,7 @@ public partial class ScreenshotCollectorPage : ComponentBase
 
     [Parameter]
     [SupplyParameterFromQuery(Name = "type")]
-    public string ScreenshotCollectionTypeString { get; set; } = "";
+    public string ScreenshotCollectionType { get; set; } = "";
 
     protected override async Task OnInitializedAsync()
     {
@@ -41,80 +37,19 @@ public partial class ScreenshotCollectorPage : ComponentBase
             )
         );
 
-        var type = Enum.Parse<ScreenshotCollectionType>(
-            ScreenshotCollectionTypeString.ToLowerInvariant()
-        );
-        switch (type)
+        switch (ScreenshotCollectionType.ToLower())
         {
-            case ScreenshotCollectionType.workoutpage:
+            case "workoutpage":
                 await HandleWorkoutPageScreenshotCollection();
+                break;
+            case "programeditor":
+                await HandleProgramEditorScreenshotCollection();
+                break;
+            case "exerciseeditor":
+                await HandleExerciseEditorScreenshotCollection();
                 break;
         }
         await base.OnInitializedAsync();
-    }
-
-    private async Task HandleWorkoutPageScreenshotCollection()
-    {
-        var demoSessionBlueprint = new SessionBlueprint(
-            Name: "Session 1",
-            Exercises:
-            [
-                new ExerciseBlueprint(
-                    Name: "Squat",
-                    Sets: 3,
-                    RepsPerSet: 10,
-                    WeightIncreaseOnSuccess: 2.5m,
-                    RestBetweenSets: Rest.Medium,
-                    SupersetWithNext: false
-                ),
-                new ExerciseBlueprint(
-                    Name: "Bench Press",
-                    Sets: 3,
-                    RepsPerSet: 10,
-                    WeightIncreaseOnSuccess: 2.5m,
-                    RestBetweenSets: Rest.Medium,
-                    SupersetWithNext: false
-                ),
-                new ExerciseBlueprint(
-                    Name: "Deadlift",
-                    Sets: 3,
-                    RepsPerSet: 10,
-                    WeightIncreaseOnSuccess: 2.5m,
-                    RestBetweenSets: Rest.Medium,
-                    SupersetWithNext: false
-                )
-            ]
-        );
-        var demoSession = demoSessionBlueprint.GetEmptySession();
-        demoSession = demoSession with
-        {
-            RecordedExercises =
-            [
-                demoSession.RecordedExercises[0] with
-                {
-                    Weight = 100m,
-                },
-                demoSession.RecordedExercises[0] with
-                {
-                    Weight = 60m,
-                },
-                demoSession.RecordedExercises[0] with
-                {
-                    Weight = 120m,
-                }
-            ],
-            Bodyweight = 85m
-        };
-        Dispatcher.Dispatch(new SetCurrentSessionAction(SessionTarget.WorkoutSession, demoSession));
-        await Task.Yield();
-        Dispatcher.Dispatch(new CycleExerciseRepsAction(SessionTarget.WorkoutSession, 0, 0));
-        Dispatcher.Dispatch(new NavigateAction("/session"));
-        await Task.Yield();
-    }
-
-    public enum ScreenshotCollectionType
-    {
-        workoutpage
     }
 }
 #endif
